% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_file}
\alias{cache_file}
\title{Get cache file path}
\usage{
cache_file(name)
}
\arguments{
\item{name}{Name of the cache file}
}
\value{
Full path to the cache file
}
\description{
Get cache file path
}
\keyword{internal}
