% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_path}
\alias{cache_path}
\title{Set up local cache directory for the package}
\usage{
cache_path()
}
\value{
Path to the cache directory
}
\description{
Set up local cache directory for the package
}
\keyword{internal}
