% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_states}
\alias{get_states}
\title{Get list of valid state codes}
\usage{
get_states()
}
\value{
A character vector of state codes
}
\description{
Returns the valid two-letter state codes that can be used with get_finance_data
}
\examples{
# Get all valid state codes
states <- get_states()
head(states)
}
