% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{is_cache_current}
\alias{is_cache_current}
\title{Check if a cached file exists and is recent}
\usage{
is_cache_current(name, max_age = 30)
}
\arguments{
\item{name}{Name of the cache file}

\item{max_age}{Maximum age in days}
}
\value{
TRUE if cache file exists and is recent, FALSE otherwise
}
\description{
Check if a cached file exists and is recent
}
\keyword{internal}
